requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
deepgram-sdk
groq
elevenlabs
edge-tts
websockets
python-multipart
